Add simulation-initialization and display of animals

+++ src/main/java/org/example/simulation/data/Animal.java
@@ -1,30 +1,30 @@
 package org.example.simulation.data;
 
 public class Animal {
-    private float x;
-    private float y;
+    private int x;
+    private int y;
     private float xVelocity;
     private float yVelocity;
     private HealthState healthState;
 
-    public Animal(float x, float y) {
+    public Animal(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public float getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public float getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(int y) {
         this.y = y;
     }
